src/requirements/celery/requirements.txt
celery
flower
redis
sqlalchemy
psycopg2-binary
python-dotenv
sqlalchemy-utils
pytz
tzlocal
requests
beautifulsoup4
playwright
